refactor(Sidebar): address PR review feedback
- Replace regex in normalizePath with a while loop for readability
- Merge two loops in resolveCurrent into one (exact + prefix match)
- Extract collection unit tests into collection.test.tsx
- Extract useSidebarNav unit tests into useSidebarNav.test.tsx
